pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-ticket-classifier-mlflow-demo"
version = "0.1.0"
description = "A GenAI ticket classification demo using MLflow GenAI and Groq."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
keywords = ["mlflow", "genai", "groq", "classification", "prompt-engineering"]

dependencies = [
    "mlflow>=2.6.0",
    "groq>=0.2.0",
    "pandas>=2.0",
    "datasets>=2.0",
    "scikit-learn>=1.2",
    "python-dotenv>=1.0",
    "click>=8.0",
    "gradio>=3.0",
    "pytest>=8.0",
]

[tool.setuptools]
packages = { find = { where = ["src"] } }

[project.urls]
"Homepage" = "https://github.com/your-org/genai-ticket-classifier-mlflow-demo"

[project.scripts]
"ticket-classifier" = "genai_ticket_classifier.cli:main"
